+++ src/main/java/samj/mario/editor/data/Palette.java
@@ -1,4 +1,4 @@
-package samj.mario.editor;
+package samj.mario.editor.data;
 
 public enum Palette {
     OVERWORLD,
