+++ src/main/java/samj/mario/editor/data/TileData.java
@@ -1,4 +1,4 @@
-package samj.mario.editor;
+package samj.mario.editor.data;
 
 import java.util.ArrayList;
 import java.util.Collections;
@@ -7,9 +7,9 @@ import java.util.Map;
 import java.util.function.Function;
 import java.util.stream.Collectors;
 
-import static samj.mario.editor.ForegroundTileCategory.INTERACTIVE;
-import static samj.mario.editor.ForegroundTileCategory.PLATFORM;
-import static samj.mario.editor.SpriteSheet.TILES;
+import static samj.mario.editor.data.ForegroundTileCategory.INTERACTIVE;
+import static samj.mario.editor.data.ForegroundTileCategory.PLATFORM;
+import static samj.mario.editor.data.SpriteSheet.TILES;
 
 public class TileData {
