+++ src/main/java/samj/mario/editor/io/LevelFormat.java
@@ -1,4 +1,6 @@
-package samj.mario.editor;
+package samj.mario.editor.io;
+
+import samj.mario.editor.data.Level;
 
 public interface LevelFormat {
     byte[] encode(Level level);
