+++ src/main/java/samj/mario/editor/io/json/JsonLevel.java
@@ -1,0 +1,4 @@
+package samj.mario.editor.io.json;
+
+public class JsonLevel {
+}
